+++ scripts/item_modifiers.py
@@ -29,17 +29,18 @@ def main(output_directory='/output'):
     maxLVL = df['MaxLVL'].tolist()
 
     if output_directory == '/output':
-        rmtree(os.path.dirname(__file__) + output_directory)
+        if os.path.exists(os.path.dirname(__file__) + output_directory):
+            rmtree(os.path.dirname(__file__) + output_directory)
         os.mkdir('output')
 
     for enchant in range(len(enchantments)):  # Loop through each enchantment file
         new_file = open(f'{output_directory.strip("/")}/{enchantments[enchant]}.json', 'w')
         new_file.truncate()  # Removes the contents of the file
-        new_file.write(f'{{\n\t\"function\": \"set_lore\",\n\t\"lore\": [{{\"text\":\"{descriptions[enchant]}\",\"italic\":false,\"color\":\"#e699e6\"}},{{\"text\":\"For: {applications[enchant]} | Max: {maxLVL[enchant]}\",\"italic\":true,\"color\":\"#bfbfbf\"}}],\n\t\"replace\": \"true\"\n}}')
+        new_file.write(f'{{\n\t\"function\": \"set_lore\",\n\t\"lore\": [{{\"text\":\"{descriptions[enchant]}\",\"italic\":false,\"color\":\"#e699e6\"}},{{\"text\":\"For: {applications[enchant]} | Max: {maxLVL[enchant]}\",\"italic\":true,\"color\":\"#bfbfbf\"}}],\n\t\"mode\": \"replace_all\"\n}}')
         new_file.close()
 
     if __name__ == '__main__':
-        print(f'[Info] Generated 39 files in {round(time() - start_time, 3)} seconds')
+        print(f'[Info] Generated {len(enchantments)} files in {round(time() - start_time, 3)} seconds')
 
 
 if __name__ == '__main__':
